pyproject.toml
[project]
name = "xn-auth"
requires-python = ">=3.12"
authors = [
    {name = "Artemiev", email = "mixartemev@gmail.com"},
]
keywords = ["starlette", "fastapi", "auth"]
description = "Auth adapter for XN-Api framework"
readme = "README.md"
license = {text = "MIT"}
dynamic = ["version"]

dependencies = [
    "fastapi",
    'pwdlib[argon2]',
    "python-jose[cryptography]",
    "xn-model",
]

[project.optional-dependencies]
dev = [
    "build",
    "python-dotenv",
    "setuptools_scm",
    "twine",
]

[project.urls]
Homepage = "https://github.com/XyncNet/x-auth"
Repository = "https://github.com/XyncNet/x-auth"

[build-system]
requires = ["setuptools>=64", "setuptools-scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["x_auth"]

[tool.setuptools_scm]
version_scheme = "python-simplified-semver" # if "feature" in `branch_name` SEMVER_MINOR++ else SEMVER_PATCH++
local_scheme = "no-local-version"

[tool.ruff]
line-length = 120